# Accounts page loading performance improvements