We called `passwd -S` 92 times! Now with `--all` it's still called twice due to the `getLogins` calls but it's already an O(n) removal improvement so \o/

Note: we might want to investigate getting rid of the `date -d` calls and replace it with one invocation? That's maybe something for a follow up.

```
before

3168 ms
3110 ms
3645 ms
5221 ms
3733 ms

-----

after the previous accounts cleanup PR

2999 ms
4445 ms
2882 ms
3801 ms
2669 ms

-------

This PR :)

790 ms
550 ms
531 ms
561 ms
440 ms
```