@martinpitt 2024-02-09
Yay! So it [doesn't work on RHEL 8](https://cockpit-logs.us-east-1.linodeobjects.com/pull-19976-20240208-220432-ee18376e-centos-8-stream-other/log.html), but we are only days ahead of forking that -- so my proposal is to just block this on dropping RHEL 8 support from main. The [ubuntu-stable failure](https://cockpit-logs.us-east-1.linodeobjects.com/pull-19976-20240208-212630-ee18376e-ubuntu-stable-other/log.html#168) looks like  a normal flake, retrying it to make sure.