@martinpitt 2024-02-09
Looks good now, but RHEL 8 tests still fail on the unexpected message: "passwd: bad argument --all: unknown option"

So yes, apparently they don't exit with 253? Perhaps you could ignore the error when the message contains "unknown option" and replace teh magic 253 value check?