@jelly 2024-02-09
> Beautiful! Tests are putting up a fight, I retried them (again).

The GC thing seems to be happening more and more :(